+++ esnextguardian.js
@@ -1,0 +1,3 @@
+// 18 September 2015
+// https://github.com/bevry/base
+module.exports = require('esnextguardian')('./esnext/lib/typechecker.js', './es5/lib/typechecker.js', require)